requirements.txt
diff-match-patch==20121119
Django==1.10.5
django-bootstrap3==9.0.0
django-braces==1.11.0
django-debug-toolbar==1.8
django-import-export==0.5.1
et-xmlfile==1.0.1
gunicorn==19.7.1
jdcal==1.3
odfpy==1.3.5
openpyxl==2.4.8
pytz==2017.2
PyYAML==3.12
sqlparse==0.2.4
tablib==0.12.1
unicodecsv==0.14.1
whitenoise==3.3.1
xlrd==1.1.0
xlwt==1.3.0
